Day 3 2015
This was great, I used CGPoint instead of making a two dimensional array or using a set of Tuples. I learned how to make it conform to Hashable and by extension (pun intended) equatable. I'm sure I could cut down on lines if I thought about it, but I think the code is pretty clean